Former-commit-id: 98f45d8feef800f2cfd02b4812ccceaf8947566b [formerly f2494c4ece65f5f51efdc87b871446151db5981d]
Former-commit-id: 2709d02f46e904b3eeab33e511ed5b3672026277

+++ app/Http/Controllers/ImportationController.php
@@ -18,6 +18,7 @@ use App\Contact;
 use App\FailRate;
 use App\Currency;
 use App\Contract;
+use PrvValidation;
 use App\Direction;
 use App\Surcharge;
 use PrvSurchargers;
@@ -5301,12 +5302,16 @@ class ImportationController extends Controller
 
     public function DestroyAccount($id){
         try{
-            $account = AccountFcl::find($id);
-            Storage::disk('FclAccount')->delete($account->namefile);
-            $account->delete();
-            return 1;
+            $contract = Contract::where('account_id',$id)->first();
+            $data = PrvValidation::ContractWithJob($contract->id);
+            if($data['bool'] == false){
+                $account = AccountFcl::find($id);
+                Storage::disk('FclAccount')->delete($account->namefile);
+                $account->delete();
+            }
+            return response()->json(['success' => 1,'jobAssociate' => $data['bool']]);
         } catch(Exception $e){
-            return 2;
+            return response()->json(['success' => 2,'jobAssociate' => false]);
         }
     }
